Refactor tunnel
