+++ cmd/http.go
@@ -17,7 +17,6 @@
 package cmd
 
 import (
-	"fmt"
 	"net/url"
 	"os"
 
@@ -101,7 +100,5 @@ func tunnelHTTP() {
 		Subdomain:          subdomain,
 		LogLevel:           logLevel,
 	}
-	fmt.Println("Use Ctrl-c to close the tunnel")
-	semaphore := tunnel.Semaphore{}
-	tunnel.StartReverseTunnel(&tunnelConfig, nil, &semaphore)
+	tunnel.StartReverseTunnel(tunnelConfig)
 }
