+++ tunnel/tunnel.go
@@ -23,33 +23,35 @@ import (
 	"os/signal"
 	"sync"
 	"syscall"
-	"time"
 
-	"github.com/cypherpunkarmory/punch/backoff"
-	"github.com/fatih/color"
 	log "github.com/sirupsen/logrus"
-	"github.com/tj/go-spin"
 	"golang.org/x/crypto/ssh"
 )
 
-type tunnelStatus struct {
-	state string
+func StartReverseTunnel(tunnelConfig ...Config) {
+	fmt.Println("Use Ctrl-c to close the tunnels")
+	semaphore := Semaphore{}
+	wg := sync.WaitGroup{}
+	wg.Add(len(tunnelConfig))
+	jumpConn, err := connectToJumpHost(&tunnelConfig[0], &semaphore)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "We could not connect to the jump host")
+		cleanup(&tunnelConfig[0])
+	}
+	for i, _ := range tunnelConfig {
+		go startReverseTunnel(jumpConn, &tunnelConfig[i], &wg, &semaphore)
+	}
+	wg.Wait()
 }
 
-const (
-	tunnelError    string = "error"
-	tunnelDone     string = "done"
-	tunnelStarting string = "starting"
-)
-
 //StartReverseTunnel Main tunneling function. Handles connections and forwarding
-func StartReverseTunnel(tunnelConfig *Config, wg *sync.WaitGroup, semaphore *Semaphore) {
+func startReverseTunnel(jumpConn *ssh.Client, tunnelConfig *Config, wg *sync.WaitGroup, semaphore *Semaphore) {
 	defer cleanup(tunnelConfig)
 
 	if wg != nil {
 		defer wg.Done()
 	}
-	listener, err := createTunnel(tunnelConfig, semaphore)
+	listener, err := createTunnel(jumpConn, tunnelConfig, semaphore)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "%s\n", err.Error())
 		return
@@ -98,153 +100,6 @@ func StartReverseTunnel(tunnelConfig *Config, wg *sync.WaitGroup, semaphore *Sem
 		}
 	}
 
-}
-
-func createTunnel(tunnelConfig *Config, semaphore *Semaphore) (net.Listener, error) {
-	tunnelCreating := tunnelStatus{tunnelStarting}
-	createCloseChannel := make(chan os.Signal)
-	signal.Notify(createCloseChannel,
-		// https://www.gnu.org/software/libc/manual/html_node/Termination-Signals.html
-		syscall.SIGTERM, // "the normal way to politely ask a program to terminate"
-		syscall.SIGINT,  // Ctrl+C
-		syscall.SIGQUIT, // Ctrl-\
-		syscall.SIGHUP,  // "terminal is disconnected"
-	)
-	defer signal.Stop(createCloseChannel)
-	go func() {
-		<-createCloseChannel
-		log.Debugf("Closing tunnel")
-		tunnelCreating.state = tunnelError
-		for !semaphore.CanRun() {
-
-		}
-		defer semaphore.Done()
-		cleanup(tunnelConfig)
-		os.Exit(0)
-	}()
-	lvl, err := log.ParseLevel(tunnelConfig.LogLevel)
-	if err != nil {
-		log.Errorf("\nLog level %s is not a valid level.", tunnelConfig.LogLevel)
-	}
-
-	log.SetLevel(lvl)
-	log.Debugf("Debug Logging activated")
-
-	var listener net.Listener
-
-	sshPort := tunnelConfig.TunnelEndpoint.SSHPort
-	// FIXME:  This should be a LUT
-	remoteEndpointPort := "3000"
-
-	if tunnelConfig.EndpointType == "https" {
-		remoteEndpointPort = "3001"
-	}
-	if tunnelConfig.EndpointType == "tcp" {
-		remoteEndpointPort = "3002"
-	}
-	var jumpServerEndpoint = Endpoint{
-		Host: tunnelConfig.ConnectionEndpoint.Hostname(),
-		Port: tunnelConfig.ConnectionEndpoint.Port(),
-	}
-
-	// remote SSH server
-	var serverEndpoint = Endpoint{
-		Host: tunnelConfig.TunnelEndpoint.IPAddress,
-		Port: sshPort,
-	}
-
-	// remote forwarding port (on remote SSH server network)
-	var remoteEndpoint = Endpoint{
-		Host: "localhost", // localhost here is the remote SSHD daemon container
-		Port: remoteEndpointPort,
-	}
-
-	privateKey, err := readPrivateKeyFile(tunnelConfig.PrivateKeyPath)
-	if err != nil {
-		return listener, err
-	}
-	hostKeyCallBack := dnsHostKeyCallback
-	if tunnelConfig.ConnectionEndpoint.Hostname() != "api.holepunch.io" {
-		fmt.Println("Ignoring hostkey")
-		hostKeyCallBack = ssh.InsecureIgnoreHostKey()
-	}
-	sshJumpConfig := &ssh.ClientConfig{
-		User: "punch",
-		Auth: []ssh.AuthMethod{
-			ssh.Password(""),
-		},
-		HostKeyCallback: hostKeyCallBack,
-		Timeout:         0,
-	}
-
-	log.Debugf("Dial into Jump Server %s", jumpServerEndpoint.String())
-	jumpConn, err := ssh.Dial("tcp", jumpServerEndpoint.String(), sshJumpConfig)
-
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error contacting the Holepunch Server.")
-		log.Debugf("%s", err)
-		return listener, err
-	}
-
-	tunnelStartingSpinner(semaphore, &tunnelCreating)
-	exponentialBackoff := backoff.NewExponentialBackOff()
-
-	// Connect to SSH remote server using serverEndpoint
-	var serverConn net.Conn
-	for {
-		serverConn, err = jumpConn.Dial("tcp", serverEndpoint.String())
-		log.Debugf("Dial into SSHD Container %s", serverEndpoint.String())
-		if err == nil {
-			tunnelCreating.state = tunnelDone
-			break
-		}
-		wait := exponentialBackoff.NextBackOff()
-		log.Debugf("Backoff Tick %s", wait.String())
-		time.Sleep(wait)
-	}
-	sshTunnelConfig := &ssh.ClientConfig{
-		User: "punch",
-		Auth: []ssh.AuthMethod{
-			privateKey,
-		},
-		//TODO: Maybe fix this. Will be rotating so dont know if possible
-		HostKeyCallback: ssh.InsecureIgnoreHostKey(),
-		Timeout:         0,
-	}
-	ncc, chans, reqs, err := ssh.NewClientConn(serverConn, serverEndpoint.String(), sshTunnelConfig)
-	if err != nil {
-		return listener, err
-	}
-	log.Debugf("SSH Connection Established via Jump %s -> %s", jumpServerEndpoint.String(), serverEndpoint.String())
-
-	sClient := ssh.NewClient(ncc, chans, reqs)
-	// Listen on remote server port
-	listener, err = sClient.Listen("tcp", remoteEndpoint.String())
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Could not open forwarding connection on remote server\n")
-		return listener, err
-	}
-	log.Debugf("Open listen port on %s", remoteEndpoint.String())
-	return listener, nil
-}
-
-func tunnelStartingSpinner(lock *Semaphore, tunnelStatus *tunnelStatus) {
-	go func() {
-		if !lock.CanRun() {
-			return
-		}
-		defer lock.Done()
-		s := spin.New()
-		for tunnelStatus.state == tunnelStarting {
-			fmt.Printf("\rStarting tunnel %s ", s.Next())
-			time.Sleep(100 * time.Millisecond)
-		}
-		if tunnelStatus.state == tunnelDone {
-			fmt.Printf("\rStarting tunnel ")
-			d := color.New(color.FgGreen, color.Bold)
-			d.Printf("✔\n")
-		}
-	}()
 }
 func cleanup(config *Config) {
 	fmt.Println("\nClosing tunnel")
